Rename bootstrap package to initializer to match its content
