+++ controllers/initializer/init_test.go
@@ -1,4 +1,4 @@
-package bootstrap
+package initializer
 
 import (
 	"context"
